pyproject.toml
[project]
name = "bug-bash-analyzer"
version = "0.1.0"
description = "AI-driven RHOAIENG bug bash analysis pipeline"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk==0.1.48",
    "flask>=3.0",
    "jsonschema>=4.20",
    "python-dotenv>=1.1",
    "requests>=2.32",
    "scipy>=1.11",
]

[tool.setuptools.packages.find]
include = ["lib*"]
